# backend/core/docs/swagger.yaml
basePath: /
definitions:
...
  handler.NotificationResponseDTO:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      link:
        type: string
      message:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  httputil.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  movies.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
      tmdb_id:
        type: integer
    type: object
  movies.Movie:
    properties:
      backdrop_url:
        type: string
      credits:
        items:
          $ref: '#/definitions/movies.MovieCredit'
        type: array
      genres:
        items:
          $ref: '#/definitions/movies.Genre'
        type: array
      id:
        type: integer
      original_language:
        type: string
      overview:
        type: string
      poster_url:
        type: string
      release_date:
        type: string
      runtime:
        type: integer
      spoken_languages:
        type: string
      status:
        type: string
      title:
        type: string
      tmdb_id:
        type: integer
    type: object
  movies.MovieCredit:
    properties:
      character:
        type: string
      id:
        type: integer
      movie_id:
        type: integer
      person:
        $ref: '#/definitions/movies.Person'
      person_id:
        type: integer
      role:
        type: string
    type: object
  movies.Person:
    properties:
      id:
        type: integer
      name:
        type: string
      profile_url:
        type: string
      tmdb_id:
        type: integer
    type: object
  tmdb.TMDBMovie:
    properties:
      id:
        type: integer
      original_language:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      release_date:
        type: string
      title:
        type: string
      vote_average:
        type: number
    type: object
  tmdb.TMDBPersonMovieCast:
    properties:
      character:
        type: string
      id:
        type: integer
      original_title:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      release_date:
        type: string
      title:
        type: string
      vote_average:
        type: number
    type: object
host: localhost:8003
info:
  contact:
    email: suporte@screek.com
    name: Suporte screeK
    url: http://www.screek.com/support
  description: API para a plataforma de cinema screeK.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: screeK API
  version: "1.0"
paths:
...
  /movies/{id}:
    get:
      consumes:
      - application/json
      description: Retorna informações do TMDB e cache local para um filme específico.
      parameters:
      - description: ID do Filme (TMDB ID)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/movies.Movie'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Detalhes completos de um filme
      tags:
      - Movies
...
  /movies/{id}/recommendations:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de filmes recomendados baseados em um filme específico.
      parameters:
      - description: ID do Filme
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tmdb.TMDBMovie'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Recomendações de filmes similares
      tags:
      - Movies
  /movies/discover:
    get:
      consumes:
      - application/json
      description: Filtra filmes por gênero e/ou ano de lançamento.
      parameters:
      - description: ID do Gênero (TMDB ID)
        in: query
        name: genre_id
        type: integer
      - description: Ano de Lançamento
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/movies.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Descoberta de filmes com filtros
      tags:
      - Movies
  /movies/search:
    get:
      consumes:
      - application/json
      description: Realiza uma busca textual no catálogo de filmes.
      parameters:
      - description: 'Termo de busca (ex: Batman)'
        in: query
        name: q
        required: true
        type: string
      - default: MOVIE
        description: Tipo de busca (MOVIE, PERSON, USER, LIST)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/movies.Movie'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Busca filmes por título
      tags:
      - Movies
  /notifications:
    get:
      description: Retorna as últimas 20 notificações (lidas e não lidas).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.NotificationResponseDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Lista notificações do usuário
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      parameters:
      - description: ID da Notificação
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      security:
      - BearerAuth: []
      summary: Marca uma notificação como lida
      tags:
      - Notifications
  /notifications/read-all:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
      security:
      - BearerAuth: []
      summary: Marca TODAS como lidas
      tags:
      - Notifications
  /people/{id}:
    get:
      consumes:
      - application/json
      description: Retorna biografia, foto e informações do TMDB via cache local.
      parameters:
      - description: ID da Pessoa (TMDB ID)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/movies.Person'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Busca detalhes de uma pessoa (ator/diretor)
      tags:
      - Movies
  /people/{id}/movies:
    get:
      consumes:
      - application/json
      description: Retorna a filmografia (créditos) de um ator ou membro da equipe.
      parameters:
      - description: ID da Pessoa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tmdb.TMDBPersonMovieCast'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      summary: Lista filmes relacionados a uma pessoa
      tags:
      - Movies
...
  /ws:
    get:
      description: Realiza o upgrade HTTP para WebSocket para receber alertas em tempo
        real.
      responses: {}
      security:
      - BearerAuth: []
      summary: Conexão WebSocket para Notificações
      tags:
      - Notifications
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"